removed universal demo components
