+++ package.json
@@ -75,21 +75,23 @@
     "accepts": "^1.3.3",
     "angular2-template-loader": "^0.4.0",
     "awesome-typescript-loader": "^2.2.4",
-    "codelyzer": "^2.0.0-beta.1",
+    "codelyzer": "1.0.0-beta.3",
     "cookie-parser": "^1.4.3",
     "express-interceptor": "^1.2.0",
     "iltorb": "^1.0.13",
     "imports-loader": "^0.6.5",
     "json-loader": "^0.5.4",
     "memory-cache": "^0.1.6",
+    "node-sass": "^3.13.0",
     "nodemon": "^1.10.0",
     "raw-loader": "^0.5.1",
     "reflect-metadata": "0.1.8",
     "rimraf": "^2.5.4",
+    "sass-loader": "^4.0.2",
     "string-replace-loader": "^1.0.5",
     "ts-helpers": "^1.1.2",
     "ts-node": "^1.3.0",
-    "tslint": "^4.0.2",
+    "tslint": "3.13.0",
     "typescript": "2.0.2",
     "v8-lazy-parse-webpack-plugin": "^0.3.0",
     "webpack": "2.1.0-beta.27",
