--- src/+app/+about/about.component.ts
@@ -1,14 +1,0 @@
-import { Component, Inject, ChangeDetectionStrategy, ViewEncapsulation } from '@angular/core';
-
-@Component({
-  changeDetection: ChangeDetectionStrategy.Default,
-  encapsulation: ViewEncapsulation.Emulated,
-  selector: 'about',
-  template: 'About component'
-})
-export class AboutComponent {
-  constructor(@Inject('req') req: any) {
-    // console.log('req',  req)
-
-  }
-}
