--- src/+app/+about/about.module.ts
@@ -1,16 +1,0 @@
-import { NgModule } from '@angular/core';
-
-import { SharedModule } from '../shared/shared.module';
-import { AboutComponent } from './about.component';
-import { AboutRoutingModule } from './about-routing.module';
-
-@NgModule({
-  imports: [
-    SharedModule,
-    AboutRoutingModule
-  ],
-  declarations: [
-    AboutComponent
-  ]
-})
-export class AboutModule { }
