--- src/+app/+home/home.component.html
@@ -1,6 +1,0 @@
-<div class="home">
-  Home component
-  <strong>Async data call return value:</strong>
-  <pre>{{ data | json }}</pre>
-  <blockquote>{{ data.data }}</blockquote>
-</div>
