--- src/+app/+lazy/lazy-routing.module.ts
@@ -1,13 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule } from '@angular/router';
-
-import { LazyComponent } from './lazy.component';
-
-@NgModule({
-  imports: [
-    RouterModule.forChild([
-      { path: '', component: LazyComponent }
-    ])
-  ]
-})
-export class LazyRoutingModule { }
