--- src/+app/+lazy/lazy.component.ts
@@ -1,14 +1,0 @@
-import { Component, Inject, ChangeDetectionStrategy, ViewEncapsulation } from '@angular/core';
-
-@Component({
-  changeDetection: ChangeDetectionStrategy.Default,
-  encapsulation: ViewEncapsulation.Emulated,
-  selector: 'lazy',
-  template: `
-    <p>
-      Lazy component
-    </p>
-  `
-})
-export class LazyComponent {
-}
