+++ src/app/app-routing.module.ts
@@ -1,13 +1,11 @@
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 
-import { AboutComponent } from './about.component';
-
 @NgModule({
   imports: [
     RouterModule.forChild([
-      { path: 'about', component: AboutComponent }
+      { path: '', redirectTo: '/home', pathMatch: 'full' }
     ])
-  ]
+  ],
 })
-export class AboutRoutingModule { }
+export class AppRoutingModule { }
