+++ src/app/app.component.html
@@ -1,0 +1,9 @@
+<h3>DSpace</h3>
+<nav>
+  <a routerLinkActive="router-link-active" routerLink="home">Home</a>
+</nav>
+<div class="container-fluid">
+  <main>
+    <router-outlet></router-outlet>
+  </main>
+</div>
