+++ src/app/app.component.scss
@@ -1,0 +1,1 @@
+
