+++ src/app/home/home.component.html
@@ -1,0 +1,3 @@
+<div class="home">
+  Home component
+</div>
