+++ src/app/home/home.module.ts
@@ -1,12 +1,10 @@
 import { NgModule } from '@angular/core';
 
-import { SharedModule } from '../shared/shared.module';
 import { HomeComponent } from './home.component';
 import { HomeRoutingModule } from './home-routing.module';
 
 @NgModule({
   imports: [
-    SharedModule,
     HomeRoutingModule
   ],
   declarations: [
