+++ src/index.html
@@ -7,16 +7,14 @@
 
   <link rel="icon" href="data:;base64,iVBORw0KGgo=">
 
-  <link rel="prerender" href="http://localhost:3000/lazy">
-  <link rel="preload" href="/assets/logo.svg">
   <base href="/">
 
 </head>
 <body>
 
-  <app>
+  <ds-app>
     Loading DSpace ...
-  </app>
+  </ds-app>
 
   <script async src="/main.bundle.js"></script>
 </body>
