+++ src/server.routes.ts
@@ -10,8 +10,5 @@
  * ];
  **/
 export const routes: string[] = [
-  'about',
-  'home',
-  'todo',
-  'lazy',
+  'home'
 ];
