+++ tslint.json
@@ -92,8 +92,12 @@
       "check-separator",
       "check-type"
     ],
-    "directive-selector": [true, "attribute", "ds", "camelCase"],
-    "component-selector": [true, "element", "ds", "kebab-case"],
+    "directive-selector-prefix": [true, "ds"],
+    "component-selector-prefix": [true, "ds"],
+    "directive-selector-name": [true, "camelCase"],
+    "component-selector-name": [true, "kebab-case"],
+    "directive-selector-type": [true, "attribute"],
+    "component-selector-type": [true, "element"],
     "use-input-property-decorator": true,
     "use-output-property-decorator": true,
     "use-host-property-decorator": true,
